+++ ra-optimisation/rao-api/src/test/java/com/powsybl/openrao/raoapi/InterTemporalRaoTest.java
@@ -18,7 +18,6 @@ import com.powsybl.openrao.commons.TemporalDataImpl;
 import com.powsybl.openrao.data.crac.api.Crac;
 import com.powsybl.openrao.data.raoresult.api.ComputationStatus;
 import com.powsybl.openrao.data.raoresult.api.RaoResult;
-import com.powsybl.openrao.raoapi.parameters.InterTemporalRao;
 import com.powsybl.openrao.raoapi.parameters.RaoParameters;
 import com.powsybl.openrao.raoapi.raomock.AnotherInterTemporalRaoProviderMock;
 import com.powsybl.openrao.raoapi.raomock.InterTemporalRaoProviderMock;
